Refactor services to base verb
